Refactor: jwt user interface 경로 변경
